Move to DataChunk extraction

+++ tools/juliapkg/src/validity_mask.jl
@@ -12,22 +12,22 @@ end
 
 const BITS_PER_VALUE = 64;
 
-function GetEntryIndex(row_idx::Int64)::Int64
+function GetEntryIndex(row_idx)
     return ((row_idx - 1) ÷ BITS_PER_VALUE) + 1
 end
 
-function GetIndexInEntry(row_idx::Int64)::Int64
+function GetIndexInEntry(row_idx)
     return (row_idx - 1) % BITS_PER_VALUE
 end
 
-function SetInvalid(mask::ValidityMask, index::Int64)
+function SetInvalid(mask::ValidityMask, index)
     entry_idx = GetEntryIndex(index)
     index_in_entry = GetIndexInEntry(index)
     mask.data[entry_idx] &= ~(1 << index_in_entry)
     return
 end
 
-function IsValid(mask::ValidityMask, index::Int64)::Bool
+function IsValid(mask::ValidityMask, index)::Bool
     entry_idx = GetEntryIndex(index)
     index_in_entry = GetIndexInEntry(index)
     return (mask.data[entry_idx] & (1 << index_in_entry)) != 0
